refactor: move StreamLogs to Pipeline interface method
